AT6394 - completed code cleanup and added icons to usa-gov banner.
