+++ src/main.ts
@@ -5,18 +5,18 @@ import "./registerServiceWorker";
 import router from "./router";
 import store from "./store";
 import vuetify from "./plugins/vuetify";
-import moment from 'moment';
-import cssVars from 'css-vars-ponyfill'; // needed to process css vars in IE11
+import moment from "moment";
+import cssVars from "css-vars-ponyfill"; // needed to process css vars in IE11
 
 import ATATFooter from "./components/ATATFooter.vue";
-import ATATSelect from "./components/ATATSelect.vue";
 import ATATHeader from "./components/ATATHeader.vue";
 import ATATHeaderNav from "./components/ATATHeaderNav.vue";
+import ATATSelect from "./components/ATATSelect.vue";
 import ATATSideBar from "./components/ATATSideBar.vue";
 import ATATTextField from "./components/ATATTextField.vue";
 import SecurityBanner from "./components/SecurityBanner.vue";
 import StyledFields from "./components/StyledFields.vue";
-import USWDCBanner from "./components/USWDCBanner.vue";
+import USAGovBanner from "./components/USAGovBanner.vue";
 
 Vue.component('atat-footer', ATATFooter);
 Vue.component('atat-header', ATATHeader);
@@ -26,7 +26,7 @@ Vue.component('atat-sidebar', ATATSideBar);
 Vue.component('atat-text-field', ATATTextField);
 Vue.component('security-banner', SecurityBanner);
 Vue.component('styled-fields', StyledFields);
-Vue.component('usdwc-banner', USWDCBanner);
+Vue.component('usa-gov-banner', USAGovBanner);
 
 Vue.config.productionTip = false;
 Vue.prototype.moment = moment;
