+++ src/sass/atat.scss
@@ -7,7 +7,7 @@
 @import "./components/app-bar-top-nav";
 
 body,html{
-    font-size: 10px !important;
+    // font-size: 10px !important;
 }
 .v-main{
     padding: 90px 0px 39px 0px !important;
