+++ src/sass/components/_app-bar-top-nav.scss
@@ -29,5 +29,102 @@
         {
             height: 0px !important;
         }
+
+        //top nav header
+        // .atat-header-nav__link {
+        //     border-radius: unset;
+        
+        //     // &.v-btn{
+        //         // padding-left: 25px;
+        //         // margin-right: 25px;
+          
+        //         // &:not(.v-btn--round).v-size--default {
+        //         //     height: 0;
+        //         // }
+
+        //         // &.v-btn--has-bg {
+        //         //     background-color: none;
+        //         // }
+
+        //         // &.atat-header-nav__support,
+        //         // &.atat-header-nav__user-display-name {
+        //         //   > .v-btn__content {
+        //         //     &::before {
+        //         //       content: " ";
+        //         //       width: 16px;
+        //         //       height: 16px;
+        //         //       margin-right: 8px;
+        //         //       background-image: url('/img/icons/avatar_white.svg');
+        //         //       color: #f0f0f0;
+        //         //       fill: #f0f0f0;
+        //         //     }
+        //         //   }
+        //         // }
+
+
+        //     //}
+
+        
+        //     // &.atat-header-nav__support,
+        //     // &.atat-header-nav__user-display-name {
+        //     //   > span:first-child {
+        //     //     display: inline-block;
+        //     //     position: relative;
+        //     //     overflow: visible;
+        //     //     color: #f0f0f0;
+        //     //     &::before {
+        //     //       content: " ";
+        //     //       width: 16px;
+        //     //       height: 16px;
+        //     //       position: absolute;
+        //     //       display: block;
+        //     //       left: -24px;
+        //     //       top: 3px;
+        //     //       margin-right: 8px;
+        //     //       background-position: top center;
+        //     //       background-size: 16px 16px;
+        //     //       background-image: url('/img/icons/avatar_white.svg');
+        //     //       color: #f0f0f0;
+        //     //       fill: #f0f0f0;
+        //     //     }
+        //     //   }
+        //     // }
+        //     // &.atat-header-nav__support {
+        //     //   margin-right: -2px;
+        //     //   > span:first-child::before {
+        //     //     background-image: url('/img/icons/help_white.svg');
+        //     //   }
+        //     // }
+        //     // &.atat-header-nav__logout {
+        //     //   &.v-btn:not(.v-btn--round).v-size--default {
+        //     //     padding-left: 0px;
+        //     //   }
+        //     //   margin-right: 25px;
+        //     //   padding-right: 25px;
+        //     //   margin-left: 10px;
+        
+        //     //   > span:first-child {
+        //     //     display: inline-block;
+        //     //     position: relative;
+        //     //     overflow: visible;
+        //     //     &::after {
+        //     //       position: absolute;
+        //     //       display: block;
+        //     //       content: " ";
+        //     //       margin-left: 8px;
+        //     //       width: 16px;
+        //     //       height: 16px;
+        //     //       right: -24px;
+        //     //       top: 3px;
+        //     //       background-position: top center;
+        //     //       background-size: 16px 16px;
+        //     //       background-image: url("/img/icons/logout_white.svg");
+        //     //       background-repeat: no-repeat;
+        //     //       color: #f0f0f0;
+        //     //       fill: #f0f0f0;
+        //     //     }
+        //     //   }
+        //     // }
+        // }
     }
 }
