+++ src/sass/components/_v-button.scss
@@ -8,6 +8,7 @@
 
   &::before{
     display: block;
+    transition: none;
   }
   
   &:hover::before{
@@ -57,6 +58,22 @@
       box-shadow: none !important;
       -ieVar-❗border-color: none!important;
     }
+  }
+
+  &.primary_darken.theme--light{
+    background-color: transparent !important;
+    color: #fff !important;
+    &:hover::before,
+    span:hover{
+     //background-color: var(--v-primary-darker-base) !important;
+     //-ieVar-❗background-color: var(--v-primary-darker-base) !important;
+      background-color: transparent;
+      // opacity: 1 !important;
+      box-shadow: none !important;
+      -ieVar-❗border-color: none!important;
+    }
+
+
   }
 
 }
